refactor user parameters to .env file
Modify the RAG with multi agents application to consume all of its arguments from a .env file, using the dotenv package, and set the default directories for the input PDF files and the output index to subdirectories of the current working directory, using relative pathing. Hide these directories and contents from git.